+++ src/app/(root)/(routes)/mypage/suggest-checks/components/my-item-summary-card/MyItemSummaryCard.tsx
@@ -1,25 +1,12 @@
-import { formatDistanceToNow } from 'date-fns'
-import koLocale from 'date-fns/locale/ko'
-import TradeStateCard from '@/components/domain/card/trade-state-card'
-import { getItemInfo } from '@/services/item/item'
+import TradeStatusCard from '@/components/domain/card/trade-status-card'
+import { Item } from '@/types'
 
-async function getItemValue(itemId: string) {
-  try {
-    const res = await getItemInfo(itemId)
-    const data = await res
-    console.log(res)
-    data.data.cardResponseDto.image =
-      'https://cdn.cetizen.com/CDN/market/market_large_crop/202203/20220318/220318152808_1_2913635.jpg'
-    console.log(data)
-    return data.data.cardResponseDto
-  } catch (e) {
-    console.log(e)
-  }
+type MyItemSummaryCardProps = {
+  item: Item
+  params: any
 }
 
-const MyItemSummaryCard = async ({ params }: { params: any }) => {
-  const data = await getItemValue('3')
-  console.log('ditto', data)
-  return <TradeStateCard item={data} className={'border-none'} />
+const MyItemSummaryCard = ({ item, params }: MyItemSummaryCardProps) => {
+  return <TradeStatusCard item={item} className={'border-none'} />
 }
 export default MyItemSummaryCard
